dev/core#772 fix php7.2 warnings on import
This is replicable if you have an import file with the column 'Contact Source' among others & import setting it to
use 'Column Headers'. It correctly maps contact source & sets it as a default. The code in setValue
is open to either a string or an array but it uses count() in it's initial check - which was OK back when it was written
- nowadays count on a string puts out an error (but still returns 1) as per originally