Remove unused SeleniumRC binaries
Overview
--------

Remove binary copy of Selenium server, because it's not used.

On a local copy of `civicrm-packages.git`, it reduces the size of the extant
files from 33mb to 13mb.

This is really just follow-up to 5.13's https://github.com/civicrm/civicrm-core/pull/13861

Hopefully, some day, Selenium or a similar E2E mechanism will rise again in
better form.  But it will probably look different
(composer/docker/npm-ified), and there's no point in carrying the `.jar`
file in here.