dev/joomla#46 - Fixed KC Finder Not Working Under Joomla 4
Updated the authentication process for Joomla 4 in KCFinder as described in GitLab (https://lab.civicrm.org/dev/joomla/-/issues/46).  The problem that needed to be fixed is that the Browse Server button will throw a 500 error under Joomla 4.  This is because Joomla changed JFactory::getApplication, so it doesn't return the same values as it did under previous versions.