Hack Smarty3+ getTemplateVars into Smarty2
Per https://lab.civicrm.org/dev/core/-/issues/4955
overriding this function in a core compatibility class allows us
to upgrade to Smarty3 & 4 but we hit a brick wall at 5.

Having it into the package we plan on removing soon gets us
past this (I will remove from the core SmartyCompatibily
once this is merged).

Not sure if there are other functions - this is the blocking one ATM